light-infosec/requirements.txt
PyCryptodome
des